# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: C4 Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /api/admin
...
paths:
...
  /pricing/sync:
    post:
      operationId: PostPricingSync
      tags: [pricing]
      summary: 手动触发一次价格同步（fetch 官方价格表 → upsert → 快照重载；失败保留旧价）
      responses:
        '200':
          description: 拉取统计
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PricingSyncResponse' }
        default:
          $ref: '#/components/responses/Error'
  /pricing/sync/preview:
    post:
      operationId: PostPricingSyncPreview
      tags: [pricing]
      summary: 预览价格同步（零写库 diff 报告）
      responses:
        '200':
          description: 预览报告
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PricingSyncPreviewResponse' }
        default:
          $ref: '#/components/responses/Error'
  /prices:
    get:
      operationId: GetPrices
      tags: [pricing]
      summary: 统一价格列表（分页+筛选）
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 1000 } }
        - { name: mode, in: query, schema: { type: string, enum: [token, call, image] } }
        - { name: source, in: query, schema: { type: string, enum: [litellm, manual] } }
        - { name: provider, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 价格列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceEntryListResponse' }
        default:
          $ref: '#/components/responses/Error'
  # 模型 ID 含 "/"（litellm 命名空间惯例 openai/gpt-5.6-sol）——一律查询参数，禁路径参数
  /prices/entry:
    get:
      operationId: GetPriceEntry
      tags: [pricing]
      summary: 单条价格查询
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 价格条目
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceEntry' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutPriceEntry
      tags: [pricing]
      summary: 手动设价（upsert）
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceEntryUpsert' }
      responses:
        '200':
          description: 更新后条目
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceEntry' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeletePriceEntry
      tags: [pricing]
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
  /prices/variants:
    get:
      operationId: GetPriceVariants
      tags: [pricing]
      summary: 变体列表（seq 升序）
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 变体列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceVariantListResponse' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutPriceVariants
      tags: [pricing]
      summary: 整体替换变体（seq 升序首匹配）
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceVariantListRequest' }
      responses:
        '200':
          description: 更新后变体列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceVariantListResponse' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeletePriceVariants
      tags: [pricing]
      summary: 清空变体
      parameters:
        - { name: model, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
  # --- /user 面（tag: user；生成到 internal/handler/user 独立 ServerInterface） ---
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean }
...
    Provider:
      type: string
      description: 主流厂商枚举（litellm_provider 实际字符串，服务前端下拉框）。**litellm_provider 动态**——litellm 更新会加新厂商，此 enum 为主流集合；新厂商出现时扩 enum（DB 筛选不受限——自由字符串等值）
      enum: [openai, anthropic, azure, vertex_ai, bedrock, deepseek, mistral, cohere, xai, openrouter, groq, together_ai, fireworks_ai, replicate, huggingface, moonshot, zhipu, baidu, alibaba, meta, nvidia, cerebras, perplexity]
    PricingSource:
      type: string
      enum: [litellm, manual]
    PricingSyncResponse:
      type: object
      required: [rows, skipped, updated]
      properties:
        rows: { type: integer, description: 拉取到的有效模型行数 }
        skipped: { type: integer, description: 解析时跳过的非法条目数（文本价、image 价与 function 价均无效，计一次） }
        updated: { type: integer, description: 文本价 upsert 落库数（manual 行不计） }
    PricingSyncPreviewResponse:
      type: object
      required: [to_add, to_update, skipped]
      properties:
        to_add: { type: integer }
        to_update: { type: integer }
        skipped: { type: integer }
        entries:
          type: array
          items: { $ref: '#/components/schemas/PricingPreviewEntry' }
        variants_changed: { type: integer }
    PricingPreviewEntry:
      type: object
      required: [model, mode, action]
      properties:
        model: { type: string }
        mode: { type: string, enum: [token, call, image] }
        action: { type: string, enum: [add, update] }
    PriceEntry:
      type: object
      required: [Model, Mode, Source, CreatedAt, UpdatedAt]
      properties:
        Model: { type: string }
        Mode: { type: string, enum: [token, call, image] }
        InputPerM: { type: number, format: double, nullable: true }
        OutputPerM: { type: number, format: double, nullable: true }
        CacheReadPerM: { type: number, format: double, nullable: true }
        CacheWritePerM: { type: number, format: double, nullable: true }
        PricePerCall: { type: number, format: double, nullable: true }
        ImgInTokPerM: { type: number, format: double, nullable: true }
        ImgOutTokPerM: { type: number, format: double, nullable: true }
        PricePerImage: { type: number, format: double, nullable: true }
        Provider: { allOf: [{ $ref: '#/components/schemas/Provider' }], nullable: true }
        Source: { $ref: '#/components/schemas/PricingSource' }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    PriceEntryUpsert:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [token, call, image] }
        input_per_m: { type: number, format: double, minimum: 0, nullable: true }
        output_per_m: { type: number, format: double, minimum: 0, nullable: true }
        cache_read_per_m: { type: number, format: double, minimum: 0, nullable: true }
        cache_write_per_m: { type: number, format: double, minimum: 0, nullable: true }
        price_per_call: { type: number, format: double, minimum: 0, nullable: true }
        img_in_tok_per_m: { type: number, format: double, minimum: 0, nullable: true }
        img_out_tok_per_m: { type: number, format: double, minimum: 0, nullable: true }
        price_per_image: { type: number, format: double, minimum: 0, nullable: true }
    PriceEntryListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/PriceEntry' }
    PriceVariant:
      type: object
      required: [Model, Seq, CreatedAt, UpdatedAt]
      properties:
        Model: { type: string }
        Seq: { type: integer }
        ServiceTier: { type: string, nullable: true }
        CtxMin: { type: integer, format: int64, nullable: true }
        CtxMax: { type: integer, format: int64, nullable: true }
        TimeStart: { type: string, nullable: true }
        TimeEnd: { type: string, nullable: true }
        DowMask: { type: integer, nullable: true }
        multiplier: { type: number, format: double, minimum: 0, maximum: 10, nullable: true, description: "价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界倍数小数——存储 mult_bp 万分：存储 15000 ↔ 显示 1.5）" }
        SetInputPerM: { type: number, format: double, nullable: true }
        SetOutputPerM: { type: number, format: double, nullable: true }
        SetCacheReadPerM: { type: number, format: double, nullable: true }
        SetCacheCreationPerM: { type: number, format: double, nullable: true }
        SetPricePerCall: { type: number, format: double, nullable: true }
        SetImgInTokPerM: { type: number, format: double, nullable: true }
        SetImgOutTokPerM: { type: number, format: double, nullable: true }
        SetPricePerImage: { type: number, format: double, nullable: true }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    PriceVariantUpsert:
      type: object
      properties:
        service_tier: { type: string, nullable: true }
        ctx_min: { type: integer, format: int64, nullable: true }
        ctx_max: { type: integer, format: int64, nullable: true }
        time_start: { type: string, nullable: true }
        time_end: { type: string, nullable: true }
        dow_mask: { type: integer, nullable: true }
        multiplier: { type: number, format: double, minimum: 0, maximum: 10, nullable: true, description: "价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界倍数小数——存储 mult_bp 万分：存储 15000 ↔ 显示 1.5）" }
        set_input_per_m: { type: number, format: double, nullable: true }
        set_output_per_m: { type: number, format: double, nullable: true }
        set_cache_read_per_m: { type: number, format: double, nullable: true }
        set_cache_creation_per_m: { type: number, format: double, nullable: true }
        set_price_per_call: { type: number, format: double, nullable: true }
        set_img_in_tok_per_m: { type: number, format: double, nullable: true }
        set_img_out_tok_per_m: { type: number, format: double, nullable: true }
        set_price_per_image: { type: number, format: double, nullable: true }
        seq: { type: integer, minimum: 1 }
    PriceVariantListResponse:
      type: object
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/PriceVariant' }
    PriceVariantListRequest:
      type: object
      properties:
        variants:
          type: array
          items: { $ref: '#/components/schemas/PriceVariantUpsert' }
...